move nav + search components
